Finished the iterative fibonacci function
Confirmed that the iterative one works. Since
it always starts with the first 2 numbers in
the sequence as 0 and 1, we establish
an array with those beginning numbers, then
check if the length number given from the user
is either 0 or 1 to just return the beginning
numbers if requested, but when it is
another number it will loop through until
it goes to the length number and add the
fibonacci numbers based on the sum of the
previous index numbers. Going to attempt
recursion for fibonacci sequence