Finished Merge sort, went step by step and fixed
When I was going through step by step, I
designed this with the pseudocode in mind
that literally just goes: 1. sort left, 2. sort right,
3. merge together. My code reflects that
where it sorts the left side first, splitting
recursively and merging when it gets down
to one element in the list, then it sorts the
right, also splitting recursively until it goes
down to one element in the list, then it merges
together the and orders the elements into a new
list. My MAIN issue was forgetting to return
the finished merged array AFTER merging.