Fixed merge sort small array bug
Since I set it so that it would go left side
first, then right side, that would mean that
if the left side had only one element
already in the array, it would initiate the
merge process anyways when I wouldn't
need to do that since that is already
considered sorted. So I made a conditional
statement to skip the merge entirely if
the given beginning array was already a
single element array and assigned it
to our "mergedLeftArr" and "mergedRightArr"
respectively.